check

+++ toeic-core-common/src/main/java/vn/myclass/core/common/utils/UploadUtils.java
@@ -5,13 +5,16 @@ import org.apache.commons.fileupload.FileItem;
 import org.apache.commons.fileupload.FileUploadException;
 import org.apache.commons.fileupload.disk.DiskFileItemFactory;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
+import org.apache.commons.lang.StringUtils;
 import org.apache.log4j.Logger;
+import vn.myclass.core.common.constant.CoreConstant;
 
 import javax.servlet.*;
 import javax.servlet.descriptor.JspConfigDescriptor;
 import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.InputStream;
+import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.*;
@@ -22,9 +25,8 @@ public class UploadUtils {
     private final Logger log = Logger.getLogger(this.getClass());
 
     public Object[] writeOrUpdateFile(HttpServletRequest request, Set<String> titleValue, String path) {
-        //Check that we have a file upload request
-        ServletContext context = request.getServletContext();
-        String adress = context.getRealPath("fileUpload");
+        String adress ="/"+ CoreConstant.FOLDER_UPLOAD;
+        checkAnhCreateFolder(adress,path);
         boolean isMultipart = ServletFileUpload.isMultipartContent(request);
         boolean check = true;
         String fileLocation = null;
@@ -51,25 +53,27 @@ public class UploadUtils {
             List<FileItem> items = upload.parseRequest(request);
             for (FileItem item : items) {
                 if (!item.isFormField()) {
-                    String fileName = item.getName();
-                    File uploadFile = new File(adress + File.separator + path + File.separator + fileName);
-                    fileLocation = adress + File.separator + path + File.separator + fileName;
-                    name = fileName;
-                    boolean isExit = uploadFile.exists();
-                    try {
-                        if(isExit){
-                            uploadFile.delete();
-                            item.write(uploadFile);
-                        }else {
-                            item.write(uploadFile);
+                    name = item.getName();
+                    if(StringUtils.isNotBlank(name)){
+                        File uploadFile = new File(adress + File.separator + path + File.separator + name);
+                        fileLocation = adress + File.separator + path + File.separator + name;
+                        boolean isExit = uploadFile.exists();
+                        try {
+                            if(isExit){
+                                uploadFile.delete();
+                                item.write(uploadFile);
+                            }else {
+                                item.write(uploadFile);
+                            }
+                        }catch (Exception e){
+                            check=false;
                         }
-                    }catch (Exception e){
-                        check=false;
                     }
+
                 } else {
                     if (titleValue != null) {
                         String nameField = item.getFieldName();
-                        String valueField = item.getString();
+                        String valueField = item.getString("UTF-8");
                         if (titleValue.contains(nameField)) {
                             mapReturnValue.put(nameField, valueField);
                         }
@@ -79,7 +83,20 @@ public class UploadUtils {
         } catch (FileUploadException e) {
             check = false;
             log.error(e.getMessage(), e);
+        } catch (UnsupportedEncodingException e) {
+            e.printStackTrace();
         }
-        return new Object[]{check, fileLocation, name, mapReturnValue};
+        return new Object[]{check, fileLocation, path+File.separator+name, mapReturnValue};
+    }
+
+    private void checkAnhCreateFolder(String adress, String path) {
+        File folderRoot=new File(adress);
+        if(!folderRoot.exists()){
+            folderRoot.mkdirs();
+        }
+        File folderChild=new File(adress+File.separator+path);
+        if(!folderChild.exists()){
+            folderChild.mkdirs();
+        }
     }
 }
